requirements.txt
discord
pymysql
datetime
async
py-dateutil
